fix: add --preserve-env=PATH to sudo command
This ensures that Nomad inherits the current user's PATH environment variable,
rather that the default sudo behavior of using a sanitized PATH.

This is necessary because the user may have installed bindle-server, traefik,
etc. locally but not system-wide. For example, cargo install installs in the
current user's home directory by default.

Signed-off-by: Joel Dice <joel.dice@gmail.com>