Added an early exit for unsupported OS environments
This is a first-time user experience improvement.

Since nomad, consul, and spin all have native Windows releases,
running the installer from a Windows git bash environment (msys) sort of works,
but the traefik job fails internally, without a clear reason in the console (other than "it took too long").

The source of the failure is buried in the nomad.log, since the traefik component download 404-ed
because the windows release is a zip, not a tar.gz.

The added OS environment check at the start of the platform installer will point n00bs in the right direction.

Signed-off-by: Doug Schmidt <doug.schmidt@aquaticinformatics.com>